test: Add async tests for ordering
